Adds multiple attachment work for Toot Views
