+++ app/src/main/java/io/github/koss/mammut/feature/instance/subfeature/feed/TootViewHolder.kt
@@ -13,11 +13,15 @@ import androidx.core.animation.doOnStart
 import androidx.core.view.*
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ViewModelProvider
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import androidx.transition.AutoTransition
 import androidx.transition.TransitionManager
+import arrow.optics.typeclasses.At
 import com.bumptech.glide.load.resource.bitmap.FitCenter
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions.withCrossFade
 import com.bumptech.glide.request.RequestOptions
+import com.github.ajalt.flexadapter.FlexAdapter
 import com.google.android.exoplayer2.ExoPlayerFactory
 import com.google.android.exoplayer2.Player
 import com.google.android.exoplayer2.SimpleExoPlayer
@@ -33,6 +37,8 @@ import io.github.koss.mammut.component.GlideApp
 import io.github.koss.mammut.data.database.entities.feed.Status
 import io.github.koss.mammut.extension.inflate
 import io.github.koss.mammut.extension.observe
+import io.github.koss.mammut.feature.instance.subfeature.feed.media.MediaAdapter
+import io.github.koss.mammut.feature.instance.subfeature.feed.media.getThumbnailSpec
 import kotlinx.android.synthetic.main.view_holder_feed_item.view.*
 import kotlinx.coroutines.*
 import org.jetbrains.anko.*
@@ -48,7 +54,6 @@ class TootViewHolder(
 ) : FeedItemViewHolder(parent.inflate(R.layout.view_holder_feed_item)), CoroutineScope by GlobalScope {
 
     private var exoPlayer: SimpleExoPlayer? = null
-
 
     private var viewModel: TootViewModel = viewModelProvider.get(UUID.randomUUID().toString(), TootViewModel::class.java)
 
@@ -78,16 +83,12 @@ class TootViewHolder(
             viewState?.content?.let(contentTextView::setText)
         }
 
-        viewState?.displayAttachment?.let(::processAttachment) ?: with(ConstraintSet()) {
-            clone(itemView.constraintLayout)
-            setDimensionRatio(itemView.tootImageCardView.id, "0")
-            applyTo(itemView.constraintLayout)
-
-            itemView.tootImageCardView.visibility = View.INVISIBLE
+        when {
+            viewState?.displayAttachments?.isEmpty() == true ->
+                itemView.attachmentsRecyclerView.isVisible = false
+            viewState?.displayAttachments?.isNotEmpty() == true ->
+                renderAttachments(viewState.displayAttachments)
         }
-
-        // Setup sensitive content screen
-        setupContentWarning(isSensitive = viewModel.currentStatus?.isSensitive ?: false)
 
         @ColorInt val color = itemView.colorAttr(R.attr.colorPrimaryLight)
 
@@ -109,161 +110,27 @@ class TootViewHolder(
         setupSpoiler(viewModel.currentStatus?.spoilerText ?: "")
     }
 
-    private fun processAttachment(attachment: Attachment<*>) {
-        with(itemView) {
-            val aspect = getThumbnailSpec(attachment)
-
+    private fun renderAttachments(attachments: List<Attachment<*>>) {
+        with (itemView) {
             with(ConstraintSet()) {
                 clone(constraintLayout)
-                setDimensionRatio(tootImageCardView.id, aspect.toString())
+                setDimensionRatio(attachmentsRecyclerView.id, (attachments.first().getThumbnailSpec() * 1.2F).toString())
                 applyTo(constraintLayout)
+
+                attachmentsRecyclerView.visibility = View.INVISIBLE
             }
 
-            tootImageCardView.visibility = View.VISIBLE
-
-            tootImageCardView.doOnLayout {
-                // Wait until the next layout pass to ensure the parent is sized correctly
-                loadAttachment(attachment)
-            }
-
-            tootImageCardView.onClick {
-                if (!viewModel.isSensitiveScreenVisible) {
-                    callbacks.onPhotoClicked(tootImageView, attachment.url)
-                }
-            }
-        }
-
-    }
-
-    private fun loadAttachment(attachment: Attachment<*>) {
-        when (attachment) {
-            is PhotoAttachment -> loadImage(attachment)
-            is VideoAttachment -> loadVideo(attachment)
-            is GifvAttachment -> loadGifv(attachment)
-        }
-    }
-
-    private fun loadGifv(gifvAttachment: Attachment<*>) {
-        itemView.tootImageView.visibility = View.GONE
-        itemView.playerView.visibility = View.VISIBLE
-        itemView.playerView.useController = false
-
-        val factory = DefaultDataSourceFactory(itemView.context,
-                Util.getUserAgent(itemView.context, "Mammut"))
-
-        val source = ExtractorMediaSource.Factory(factory)
-                .createMediaSource(Uri.parse(gifvAttachment.url))
-
-        exoPlayer = ExoPlayerFactory.newSimpleInstance(itemView.context).apply {
-            repeatMode = Player.REPEAT_MODE_ALL
-            playWhenReady = true
-        }
-
-        itemView.playerView.player = exoPlayer
-        exoPlayer?.prepare(source)
-    }
-
-    private fun loadVideo(videoAttachment: Attachment<*>) {
-        itemView.tootImageView.visibility = View.GONE
-        itemView.playerView.visibility = View.VISIBLE
-        itemView.playerView.useController = true
-
-        val factory = DefaultDataSourceFactory(itemView.context,
-                Util.getUserAgent(itemView.context, "Mammut"))
-
-        val source = ExtractorMediaSource.Factory(factory)
-                .createMediaSource(Uri.parse(videoAttachment.url))
-
-        exoPlayer = ExoPlayerFactory.newSimpleInstance(itemView.context)
-
-        itemView.playerView.player = exoPlayer
-        exoPlayer?.prepare(source)
-    }
-
-    private fun loadImage(photoAttachment: PhotoAttachment) {
-        itemView.tootImageView.visibility = View.VISIBLE
-        itemView.playerView.visibility = View.GONE
-
-        // Resolve colors
-        val typedValue = TypedValue()
-        val theme = itemView.context.theme ?: return
-        theme.resolveAttribute(R.attr.colorPrimaryLight, typedValue, true)
-        @ColorInt val color = typedValue.data
-
-        val requestManager = GlideApp.with(itemView)
-
-        // Load attachment
-        requestManager
-                .load(photoAttachment.url)
-                .thumbnail(
-                        requestManager
-                                .load(photoAttachment.previewUrl)
-                                .thumbnail(
-                                        requestManager
-                                                .load(ColorDrawable(color))
-                                )
-                                .transition(withCrossFade())
-                )
-                .transition(withCrossFade())
-                .apply(RequestOptions.bitmapTransform(FitCenter()))
-                .into(itemView.tootImageView)
-    }
-
-    private fun setupContentWarning(isSensitive: Boolean) {
-        with(itemView) {
-            // If not sensitive content, short circuit
-            if (!isSensitive) {
-                viewModel.isSensitiveScreenVisible = false
-                sensitiveContentFrameLayout.isVisible = false
-                sensitiveContentToggleButton.isVisible = false
-                return
-            }
-
-            // Initial conditions
-            sensitiveContentFrameLayout.isVisible = true
-            sensitiveContentToggleButton.isVisible = true
-            viewModel.isSensitiveScreenVisible = true
-
-            fun View.largestDimension(): Float = sqrt(this.width.toFloat().pow(2F) + this.height.toFloat().pow(2F))
-
-            fun toggleContentWarningVisibility() {
-                if (viewModel.isSensitiveScreenVisible) {
-                    ViewAnimationUtils.createCircularReveal(
-                            sensitiveContentFrameLayout,
-                            sensitiveContentFrameLayout.width - sensitiveContentToggleButton.width / 2,
-                            sensitiveContentToggleButton.height / 2,
-                            sensitiveContentFrameLayout.largestDimension(),
-                            0F
-                    ).apply {
-                        doOnEnd {
-                            sensitiveContentFrameLayout.isVisible = false
-                        }
-                        duration = 250L
-                    }.start()
-
-                    sensitiveContentToggleButton.imageResource = R.drawable.ic_visibility_black_24dp
-                    viewModel.isSensitiveScreenVisible = false
-                } else {
-                    ViewAnimationUtils.createCircularReveal(
-                            sensitiveContentFrameLayout,
-                            sensitiveContentFrameLayout.width - sensitiveContentToggleButton.width / 2,
-                            sensitiveContentToggleButton.height / 2,
-                            0F,
-                            sensitiveContentFrameLayout.largestDimension()
-                    ).apply {
-                        doOnStart {
-                            sensitiveContentFrameLayout.isVisible = true
-                        }
-                        duration = 250L
-                    }.start()
-
-                    sensitiveContentToggleButton.imageResource = R.drawable.ic_visibility_off_black_24dp
-                    viewModel.isSensitiveScreenVisible = true
+            when {
+                attachmentsRecyclerView.adapter == null -> {
+                    attachmentsRecyclerView.adapter = MediaAdapter(callbacks)
+                    attachmentsRecyclerView.layoutManager = LinearLayoutManager(context, RecyclerView.HORIZONTAL, false)
                 }
             }
 
-            sensitiveContentToggleButton.onClick { toggleContentWarningVisibility() }
-            sensitiveContentFrameLayout.onClick { toggleContentWarningVisibility() }
+            (attachmentsRecyclerView.adapter as MediaAdapter).apply {
+                contentIsSensitive = viewModel.currentStatus?.isSensitive ?: false
+                submitList(attachments)
+            }
         }
     }
 
@@ -280,8 +147,8 @@ class TootViewHolder(
 
             // Change visibility
             contentTextView.isVisible = viewModel.isContentVisible
-            val hideImageView = viewModel.statusViewState.value?.displayAttachment != null
-            if (hideImageView) tootImageCardView.isVisible = viewModel.isContentVisible
+            val hideImageView = viewModel.statusViewState.value?.displayAttachments?.isNotEmpty() == true
+            if (hideImageView) attachmentsRecyclerView.isVisible = viewModel.isContentVisible
 
             // Change button icon
             contentWarningVisibilityButton.imageResource =
@@ -296,30 +163,6 @@ class TootViewHolder(
         }
 
         renderContentVisibility(false)
-    }
-
-    /**
-     * Function for inspecting an attachment for metadata and retrieving an approximate
-     * width and height. Will assume 4:3 width:height ratio if nothing found
-     */
-    private fun getThumbnailSpec(attachment: Attachment<*>): Float {
-        val bestGuess = 400F / 300F
-        return when (attachment) {
-            is PhotoAttachment -> attachment.metadata?.original?.run {
-                when {
-                    aspect != 0F -> aspect
-                    width != 0 && height != 0 -> width.toFloat() / height.toFloat()
-                    else -> bestGuess
-                }
-            } ?: bestGuess
-            is VideoAttachment -> attachment.metadata?.original?.run {
-                if (width != 0 && height != 0) width.toFloat() / height.toFloat() else bestGuess
-            } ?: bestGuess
-            is GifvAttachment -> attachment.metadata?.original?.run {
-                if (width != 0 && height != 0) width.toFloat() / height.toFloat() else bestGuess
-            } ?: bestGuess
-            else -> throw IllegalArgumentException("Unknown attachment type")
-        }
     }
 
     fun recycle() {
